Migrate account_group_invoice_lines to version 9
