+++ account_group_invoice_lines/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import account
+from . import models
