+++ account_group_invoice_lines/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import account_journal
+from . import account_invoice
